+++ mobile/lib/core/network/network_providers.g.dart
@@ -1,0 +1,40 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'network_providers.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$tokenStorageHash() => r'b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3';
+
+/// See also [tokenStorage].
+@ProviderFor(tokenStorage)
+final tokenStorageProvider = AutoDisposeProvider<TokenStorage>.internal(
+  tokenStorage,
+  name: r'tokenStorageProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$tokenStorageHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef TokenStorageRef = AutoDisposeProviderRef<TokenStorage>;
+
+String _$apiClientHash() => r'c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4';
+
+/// See also [apiClient].
+@ProviderFor(apiClient)
+final apiClientProvider = AutoDisposeProvider<ApiClient>.internal(
+  apiClient,
+  name: r'apiClientProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$apiClientHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef ApiClientRef = AutoDisposeProviderRef<ApiClient>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package